pyproject.toml
[project]
name = "impact-engine-orchestrator"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = ["pyyaml"]

[project.optional-dependencies]
dev = ["pytest", "nbmake", "nbconvert", "jupyter", "ruff"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["impact_engine_orchestrator"]

[tool.hatch.envs.default]
features = ["dev"]

[tool.ruff]
extend-exclude = ["_external"]
line-length = 120

[tool.ruff.lint]
select = ["D", "E", "F", "I"]

[tool.ruff.lint.pydocstyle]
convention = "numpy"

[tool.ruff.lint.per-file-ignores]
"*.ipynb" = ["D", "E402", "F", "I"]
"tests/**" = ["D"]
"scripts/**" = ["D"]

[tool.pytest.ini_options]
addopts = "-v"
